refactor(user): 修改用户ID字段类型为Integer
- 将用户模型中的id字段从long类型改为Integer类型
- 更新getId方法返回类型为Integer
- 更新setId方法参数类型为Integer

+++ src/main/java/week4/app/models/User.java
@@ -11,7 +11,7 @@ public class User {
     /**
      * ID 编号
      */
-    private long id;
+    private Integer id;
     /**
      * 密码，用户的密码
      */
@@ -29,11 +29,11 @@ public class User {
         this.createTime = value;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long value) {
+    public void setId(Integer value) {
         this.id = value;
     }
